+++ src/spring/azext_spring/vendored_sdks/appplatform/v2022_09_01_preview/operations/__init__.py
@@ -31,7 +31,7 @@ from ._api_portals_operations import ApiPortalsOperations
 from ._api_portal_custom_domains_operations import ApiPortalCustomDomainsOperations
 
 from ._patch import __all__ as _patch_all
-from ._patch import *  # type: ignore # pylint: disable=unused-wildcard-import
+from ._patch import *  # pylint: disable=unused-wildcard-import
 from ._patch import patch_sdk as _patch_sdk
 
 __all__ = [