+++ src/containerapp/HISTORY.rst
@@ -4,6 +4,8 @@ Release History
 ===============
 Upcoming
 +++++++
+* 'az containerapp env update': Fix bugs in update environment.
+* Fix YAML create with user-assigned identity
 * Fix polling logic for long running operations.
 * 'az containerapp env create': add support for workload profiles
 * 'az containerapp env update': add support for workload profiles